# schemas/openapi-v1alpha1.json
{
  "openapi": "3.1.0",
  "info": {"title":"DARKSTAR local API","version":"1.0.0-alpha.1"},
  "servers": [{"url":"http://127.0.0.1:{port}","variables":{"port":{"default":"0"}}}],
  "security": [{"bearerAuth":[]}],
  "paths": {
    "/api/v1/work-items/source-views":{"get":{"operationId":"listWorkSourceViews","parameters":[{"name":"projectId","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Version-bound source admission and independent execution facts","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkSourceViews"}}}},"default":{"$ref":"#/components/responses/Error"}}}},
...
    "/api/v1/projects/{projectId}/backlog/executions":{"get":{"operationId":"listTicketExecutions","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"observationId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Version-bound source admission and independent execution facts","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkSourceViews"}}}},"default":{"$ref":"#/components/responses/Error"}}}},
...
    "/api/v1/projects/{projectId}/tickets":{"get":{"operationId":"listNativeTickets","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"q","in":"query","schema":{"type":"string"}},{"name":"state","in":"query","schema":{"type":"string"}},{"name":"cursor","in":"query","schema":{"type":"string"}},{"name":"pageSize","in":"query","schema":{"type":"integer","minimum":1,"maximum":100}}],"responses":{"200":{"description":"Native business tickets with preserved history","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NativeTicketPage"}}}},"default":{"$ref":"#/components/responses/Error"}}}},
    "/api/v1/projects/{projectId}/tickets/{ticketId}":{"get":{"operationId":"getNativeTicket","parameters":[{"name":"projectId","in":"path","required":true,"schema":{"type":"string"}},{"name":"ticketId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Native business tickets with preserved history","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NativeTicketDetail"}}}},"default":{"$ref":"#/components/responses/Error"}}}},
...
    "/api/v1/projects": {
      "get": {"operationId":"listProjects","responses":{"200":{"description":"Registered project projections","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Project"}}}}},"default":{"$ref":"#/components/responses/Error"}}},
      "post": {"operationId":"registerProject","parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectRegistration"}}}},"responses":{"201":{"description":"Project registered","headers":{"Location":{"schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"default":{"$ref":"#/components/responses/Error"}}}
    },
...
    "/api/v1/work-items": {
      "get": {"operationId":"listWorkItems","parameters":[{"name":"includeDeleted","in":"query","required":false,"schema":{"type":"boolean","default":false}},{"name":"projectId","in":"query","required":false,"schema":{"type":"string","pattern":"^project_[0-9A-HJKMNP-TV-Z]{26}$"}}],"responses":{"200":{"description":"Work-item projections","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkItem"}}}}},"default":{"$ref":"#/components/responses/Error"}}},
      "post": {"operationId":"createWorkItem","parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWorkItemRequest"}}}},"responses":{"201":{"description":"Authored work item created","headers":{"Location":{"schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkItem"}}}},"default":{"$ref":"#/components/responses/Error"}}}
    },
...
  },
  "components": {
    "securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer"}},
    "parameters": {
...
      "IdempotencyKey":{"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string","minLength":8,"maxLength":128}},
...
    },
    "responses":{"Error":{"description":"Stable API error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}},
    "schemas": {
...
      "SourceTicketObservation":{"type":"object","additionalProperties":false,"required":["observationId","ref","revision","title","description","key","url","businessState","observedAt","evidenceRef"],"properties":{"observationId":{"type":"string"},"ref":{"$ref":"#/components/schemas/BacklogTicketRef"},"revision":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"key":{"type":"string"},"url":{"type":"string"},"businessState":{"$ref":"#/components/schemas/BacklogNamedObservation"},"observedAt":{"type":"string"},"evidenceRef":{"type":"string"}}},
      "SourceLineage":{"type":"object","additionalProperties":false,"required":["workItemId","projectId","ticketKey","revision","bindingRevision","ref","origin","createdAt"],"properties":{"workItemId":{"type":"string"},"projectId":{"type":"string"},"ticketKey":{"type":"string"},"revision":{"type":"integer","minimum":1},"bindingRevision":{"type":"integer","minimum":1},"ref":{"$ref":"#/components/schemas/BacklogTicketRef"},"origin":{"enum":["admitted","native","legacy_native"]},"createdAt":{"type":"string"}}},
      "TicketAdmission":{"type":"object","additionalProperties":false,"required":["id","workItemId","projectId","observationId","lineageRevision","bindingRevision","approvedAt","actor"],"properties":{"id":{"type":"string"},"workItemId":{"type":"string"},"projectId":{"type":"string"},"observationId":{"type":"string"},"lineageRevision":{"type":"integer","minimum":1},"bindingRevision":{"type":"integer","minimum":1},"approvedAt":{"type":"string"},"actor":{"type":"string"}}},
      "SourceAssessment":{"type":"object","additionalProperties":false,"required":["state","reasons"],"properties":{"state":{"enum":["ready","action_required","unresolved_source"]},"reasons":{"type":"array","items":{"type":"string"}}}},
      "StringObservation":{"oneOf":[{"type":"object","additionalProperties":false,"required":["state","value"],"properties":{"state":{"const":"known"},"value":{"type":"string"}}},{"type":"object","additionalProperties":false,"required":["state","reason"],"properties":{"state":{"const":"unknown"},"reason":{"type":"string"}}},{"type":"object","additionalProperties":false,"required":["state","reason"],"properties":{"state":{"const":"unsupported"},"reason":{"type":"string"}}}]},
      "WorkSourceView":{"type":"object","additionalProperties":false,"required":["schemaVersion","workItemId","projectId","localActivity","runOutcome","approvedTicket","currentTicket","currentObservationId","lineage","lineages","approval","assessment","runs","externalAcceptance"],"properties":{"schemaVersion":{"const":1},"workItemId":{"type":"string"},"projectId":{"type":"string"},"localActivity":{"enum":["idle","draft","ready","queued","running","waiting","blocked","failed","reconcile_required"]},"runOutcome":{"enum":["unobserved","completed","failed","cancelled"]},"approvedTicket":{"anyOf":[{"$ref":"#/components/schemas/SourceTicketObservation"},{"type":"null"}]},"currentTicket":{"anyOf":[{"$ref":"#/components/schemas/SourceTicketObservation"},{"type":"null"}]},"currentObservationId":{"type":"string"},"lineage":{"anyOf":[{"$ref":"#/components/schemas/SourceLineage"},{"type":"null"}]},"lineages":{"type":"array","items":{"$ref":"#/components/schemas/SourceLineage"}},"approval":{"anyOf":[{"$ref":"#/components/schemas/TicketAdmission"},{"type":"null"}]},"assessment":{"$ref":"#/components/schemas/SourceAssessment"},"runs":{"type":"array","items":{"$ref":"#/components/schemas/Run"}},"externalAcceptance":{"$ref":"#/components/schemas/StringObservation"}}},
      "WorkSourceViews":{"type":"object","additionalProperties":false,"required":["schemaVersion","items"],"properties":{"schemaVersion":{"const":1},"items":{"type":"array","items":{"$ref":"#/components/schemas/WorkSourceView"}}}},
...
      "BacklogNamespace":{"type":"object","additionalProperties":false,"required":["provider","host","tenantId","scopeId"],"properties":{"provider":{"type":"string"},"host":{"type":"string"},"tenantId":{"type":"string"},"scopeId":{"type":"string"}}},
...
      "BacklogTicketRef":{"type":"object","additionalProperties":false,"required":["namespace","id"],"properties":{"namespace":{"$ref":"#/components/schemas/BacklogNamespace"},"id":{"type":"string"}}},
...
      "BacklogNamedID":{"type":"object","additionalProperties":false,"required":["id","name"],"properties":{"id":{"type":"string"},"name":{"type":"string"}}},
      "BacklogNamedObservation":{"oneOf":[{"type":"object","additionalProperties":false,"required":["state","value"],"properties":{"state":{"const":"known"},"value":{"$ref":"#/components/schemas/BacklogNamedID"}}},{"type":"object","additionalProperties":false,"required":["state","reason"],"properties":{"state":{"const":"unknown"},"reason":{"type":"string"}}},{"type":"object","additionalProperties":false,"required":["state","reason"],"properties":{"state":{"const":"unsupported"},"reason":{"type":"string"}}}]},
...
      "NativeTicketNamedID":{"type":"object","additionalProperties":false,"required":["id","name"],"properties":{"id":{"type":"string"},"name":{"type":"string"}}},
      "NativeTicket":{"type":"object","additionalProperties":false,"required":["id","projectId","revision","key","url","title","description","businessState","priority","assignees","labels","evidenceRef","observedAt"],"properties":{"id":{"type":"string"},"projectId":{"type":"string"},"revision":{"type":"string"},"key":{"type":"string"},"url":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"businessState":{"$ref":"#/components/schemas/NativeTicketNamedID"},"priority":{"type":"integer","minimum":0},"assignees":{"type":"array","items":{"$ref":"#/components/schemas/NativeTicketNamedID"}},"labels":{"type":"array","items":{"$ref":"#/components/schemas/NativeTicketNamedID"}},"evidenceRef":{"type":"string"},"observedAt":{"type":"string","format":"date-time"}}},
      "NativeTicketPage":{"type":"object","additionalProperties":false,"required":["schemaVersion","tickets","nextCursor"],"properties":{"schemaVersion":{"const":1},"tickets":{"type":"array","items":{"$ref":"#/components/schemas/NativeTicket"}},"nextCursor":{"type":"string"}}},
      "NativeTicketCapabilities":{"type":"object","additionalProperties":false,"required":["edit","transitions"],"properties":{"edit":{"type":"boolean"},"transitions":{"type":"boolean"}}},
      "NativeTicketField":{"type":"object","additionalProperties":false,"required":["id","name","kind","required"],"properties":{"id":{"type":"string"},"name":{"type":"string"},"kind":{"enum":["text","number","ids"]},"required":{"type":"boolean"}}},
      "NativeTicketTransition":{"type":"object","additionalProperties":false,"required":["id","name","toState"],"properties":{"id":{"type":"string"},"name":{"type":"string"},"toState":{"$ref":"#/components/schemas/NativeTicketNamedID"}}},
      "NativeTicketHistory":{"type":"object","additionalProperties":false,"required":["revision","kind","recordedAt","evidenceRef","title","description","businessState","priority"],"properties":{"revision":{"type":"string"},"kind":{"type":"string"},"recordedAt":{"type":"string","format":"date-time"},"evidenceRef":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"businessState":{"type":"string"},"priority":{"type":"integer","minimum":0}}},
      "NativeTicketDetail":{"type":"object","additionalProperties":false,"required":["schemaVersion","ticket","capabilities","fields","transitions","history"],"properties":{"schemaVersion":{"const":1},"ticket":{"$ref":"#/components/schemas/NativeTicket"},"capabilities":{"$ref":"#/components/schemas/NativeTicketCapabilities"},"fields":{"type":"array","items":{"$ref":"#/components/schemas/NativeTicketField"}},"transitions":{"type":"array","items":{"$ref":"#/components/schemas/NativeTicketTransition"}},"history":{"type":"array","items":{"$ref":"#/components/schemas/NativeTicketHistory"}}}},
...
      "ApiError":{"$ref":"./api-error-v1alpha1.schema.json"},
...
      "ProjectRegistration":{"type":"object","additionalProperties":false,"required":["name","source"],"properties":{"name":{"type":"string","minLength":1},"source":{"type":"string","minLength":1,"description":"Existing absolute project directory visible to the local daemon."}}},
      "Project":{"type":"object","additionalProperties":false,"required":["id","name","sourceHash","status","resourceVersion","lastGlobalPosition","createdAt","updatedAt"],"properties":{"id":{"type":"string","pattern":"^project_[0-9A-HJKMNP-TV-Z]{26}$"},"name":{"type":"string","minLength":1},"sourceHash":{"type":"string","pattern":"^[0-9a-f]{64}$"},"status":{"enum":["active","archived"]},"resourceVersion":{"type":"integer","minimum":1},"lastGlobalPosition":{"type":"integer","minimum":1},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},
...
      "CreateWorkItemRequest":{"type":"object","additionalProperties":false,"required":["projectId","title"],"properties":{"projectId":{"type":"string","pattern":"^project_[0-9A-HJKMNP-TV-Z]{26}$"},"title":{"type":"string","minLength":1},"details":{"type":"string"},"evidence":{"type":"array","uniqueItems":true,"items":{"type":"string","minLength":1}},"routingIntent":{"$ref":"#/components/schemas/WorkRoutingIntent"},"priority":{"type":"integer","minimum":0,"default":0,"deprecated":true}}},
      "WorkRoutingIntent":{"oneOf":[{"$ref":"#/components/schemas/AutomaticWorkRoutingIntent"},{"$ref":"#/components/schemas/OverrideWorkRoutingIntent"}]},
      "AutomaticWorkRoutingIntent":{"type":"object","additionalProperties":false,"required":["mode"],"properties":{"mode":{"const":"automatic"}}},
      "OverrideWorkRoutingIntent":{"type":"object","additionalProperties":false,"required":["mode","workflowId"],"properties":{"mode":{"const":"override"},"workflowId":{"type":"string","minLength":1},"workflowVersion":{"type":"string","minLength":1},"entryNodeId":{"type":"string","minLength":1},"terminalNodeIds":{"type":"array","uniqueItems":true,"items":{"type":"string","minLength":1}}}},
...
      "WorkItem":{"type":"object","additionalProperties":false,"required":["id","projectId","title","evidence","routingIntent","sourceHash","priority","status","resourceVersion","lastGlobalPosition","createdAt","updatedAt"],"properties":{"deletion":{"enum":["deleting","deleted"]},"id":{"type":"string","pattern":"^work_[0-9A-HJKMNP-TV-Z]{26}$"},"projectId":{"type":"string","pattern":"^project_[0-9A-HJKMNP-TV-Z]{26}$"},"title":{"type":"string","minLength":1},"details":{"type":"string"},"evidence":{"type":"array","uniqueItems":true,"items":{"type":"string","minLength":1}},"routingIntent":{"$ref":"#/components/schemas/WorkRoutingIntent"},"sourceHash":{"type":"string","pattern":"^[0-9a-f]{64}$"},"priority":{"type":"integer","minimum":0},"status":{"enum":["open","active","completed","cancelled"]},"resourceVersion":{"type":"integer","minimum":1},"lastGlobalPosition":{"type":"integer","minimum":1},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},
...
      "WorkflowDocument":{"oneOf":[{"$ref":"./workflow-v1alpha1.schema.json"},{"$ref":"./workflow-v1alpha2.schema.json"},{"$ref":"./workflow-v1alpha3.schema.json"}]},
...
      "WorkflowValidationIssue":{"type":"object","additionalProperties":false,"required":["code","message"],"properties":{"code":{"type":"string"},"message":{"type":"string"},"location":{"type":"string"},"details":{"type":"object","additionalProperties":{"type":"array","items":{"type":"string"}}}}},
...
      "PreparationQuestion":{"type":"object","additionalProperties":false,"required":["id","prompt"],"properties":{"id":{"type":"string"},"prompt":{"type":"string"}}},
      "PreparationCandidateAdvice":{"type":"object","additionalProperties":false,"required":["entry","disposition","rationale","questions","assumptions","terminals"],"properties":{"entry":{"type":"string"},"disposition":{"enum":["suitable","unsuitable","input_required"]},"rationale":{"type":"string"},"questions":{"type":["array","null"],"items":{"$ref":"#/components/schemas/PreparationQuestion"}},"assumptions":{"type":["array","null"],"items":{"type":"string"}},"terminals":{"type":"array","minItems":1,"uniqueItems":true,"items":{"type":"string"}}}},
      "PreparationAdvice":{"type":"object","additionalProperties":false,"required":["confidence","candidates","evidenceUsed"],"properties":{"confidence":{"enum":["high","medium","low"]},"candidates":{"type":["array","null"],"items":{"$ref":"#/components/schemas/PreparationCandidateAdvice"}},"evidenceUsed":{"type":["array","null"],"items":{"type":"string"}}}},
      "PreparationPolicy":{"type":"object","additionalProperties":false,"required":["version","requiredNodes","consequentialNodes","allowedAssumptions"],"properties":{"version":{"type":"string"},"requiredNodes":{"type":["array","null"],"items":{"type":"string"}},"consequentialNodes":{"type":["array","null"],"items":{"type":"string"}},"allowedAssumptions":{"type":["array","null"],"items":{"type":"string"}}}},
      "PreparationEvidence":{"type":"object","additionalProperties":false,"required":["reference"],"properties":{"reference":{"type":"string"},"digest":{"type":"string","pattern":"^[0-9a-f]{64}$"},"content":{"type":"string"}}},
      "PreparationContext":{"type":"object","additionalProperties":false,"required":[],"properties":{"runInputs":{"type":"object","additionalProperties":{}},"acceptedOutputs":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{}}},"requiredNodes":{"type":["array","null"],"items":{"type":"string"}}}},
      "PreparationInputSnapshot":{"type":"object","additionalProperties":false,"required":["work","project","workflow","workflowDigest","policy","context","answers","evidence"],"properties":{"work":{"$ref":"#/components/schemas/WorkItem"},"project":{"$ref":"#/components/schemas/Project"},"workflow":{"$ref":"#/components/schemas/WorkflowDocument"},"workflowDigest":{"type":"string","pattern":"^[0-9a-f]{64}$"},"policy":{"$ref":"#/components/schemas/PreparationPolicy"},"context":{"$ref":"#/components/schemas/PreparationContext"},"answers":{"type":["object","null"],"additionalProperties":{"type":"string"}},"evidence":{"type":["array","null"],"items":{"$ref":"#/components/schemas/PreparationEvidence"}},"override":{"type":"object","additionalProperties":false,"required":[],"properties":{"from":{"type":"string"},"until":{"type":["array","null"],"items":{"type":"string"}}}}}},
      "PreparationAlternative":{"type":"object","additionalProperties":false,"required":["entry","nodeCount","rationale","validation","missing","terminals"],"properties":{"entry":{"type":"string"},"nodeCount":{"type":"integer","minimum":0},"rationale":{"type":"string"},"validation":{"type":["array","null"],"items":{"$ref":"#/components/schemas/WorkflowValidationIssue"}},"missing":{"type":["array","null"],"items":{"$ref":"#/components/schemas/FrozenRouteInputRequirement"}},"terminals":{"type":"array","minItems":1,"uniqueItems":true,"items":{"type":"string"}}}},
      "RoutePreparationAssessment":{"type":"object","additionalProperties":false,"required":["schemaVersion","input","inputDigest","advice","route","rationale","questions","confirmationReasons","alternatives","digest"],"properties":{"schemaVersion":{"const":1},"input":{"$ref":"#/components/schemas/PreparationInputSnapshot"},"inputDigest":{"type":"string","pattern":"^[0-9a-f]{64}$"},"advice":{"$ref":"#/components/schemas/PreparationAdvice"},"route":{"$ref":"#/components/schemas/FrozenRoute"},"rationale":{"type":"string"},"questions":{"type":["array","null"],"items":{"$ref":"#/components/schemas/PreparationQuestion"}},"confirmationReasons":{"type":["array","null"],"items":{"type":"string"}},"alternatives":{"type":["array","null"],"items":{"$ref":"#/components/schemas/PreparationAlternative"}},"digest":{"type":"string","pattern":"^[0-9a-f]{64}$"}}},
...
      "FrozenRouteNode":{"type":"object","additionalProperties":false,"required":["id"],"properties":{"id":{"type":"string"},"join":{"type":"object"}}},
      "FrozenRouteTransition":{"type":"object","additionalProperties":false,"required":["id","from","to"],"properties":{"id":{"type":"string"},"from":{"type":"string"},"to":{"type":"string"}}},
      "FrozenRouteExcludedNode":{"type":"object","additionalProperties":false,"required":["id","reason"],"properties":{"id":{"type":"string"},"reason":{"enum":["before_entry","past_terminal","not_connected"]}}},
      "FrozenRouteInputRequirement":{"type":"object","additionalProperties":false,"required":["code","node","input","source"],"properties":{"code":{"type":"string"},"node":{"type":"string"},"input":{"type":"string"},"source":{"type":"string"}}},
      "FrozenRoute":{"type":"object","additionalProperties":false,"required":["entry","terminals","nodes","transitions","excludedNodes","inputRequirements"],"properties":{"entry":{"type":"string"},"terminals":{"type":"array","items":{"type":"string"}},"nodes":{"type":"array","items":{"$ref":"#/components/schemas/FrozenRouteNode"}},"transitions":{"type":"array","items":{"$ref":"#/components/schemas/FrozenRouteTransition"}},"excludedNodes":{"type":"array","items":{"$ref":"#/components/schemas/FrozenRouteExcludedNode"}},"inputRequirements":{"type":"array","items":{"$ref":"#/components/schemas/FrozenRouteInputRequirement"}},"assessment":{"$ref":"#/components/schemas/RoutePreparationAssessment"}}},
      "Run":{"type":"object","additionalProperties":false,"required":["id","workItemId","workflowId","workflowVersion","status","resourceVersion","createdAt","updatedAt"],"properties":{"id":{"type":"string","pattern":"^run_[0-9A-HJKMNP-TV-Z]{26}$"},"workItemId":{"type":"string"},"workflowId":{"type":"string"},"workflowVersion":{"type":"string"},"workflowDigest":{"type":"string","pattern":"^[0-9a-f]{64}$"},"routeDigest":{"type":"string","pattern":"^[0-9a-f]{64}$"},"routeSnapshot":{"$ref":"#/components/schemas/FrozenRoute"},"priority":{"type":"integer","minimum":0},"status":{"enum":["pending","draft","ready","queued","running","waiting","blocked","completed","failed","cancelled","reconcile_required"]},"resourceVersion":{"type":"integer","minimum":1},"lastGlobalPosition":{"type":"integer","minimum":1},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},
...
    }
  }
}